Hotfix/walls+bugs (#756)
* Removed compiler warnings in mesh.cpp

* Fixed warnings from constructor lists not
mathcing header order.

* Fixed warnings from unused functions, changed
to inline functions.

* Fixed missing braces warnings

* Fixed unused variable warnings.

* Fixed unitialized variables and misleading indentation
wrnings.

* Made MeshBoundaryValues class destructor virtual

* Some final last changes to remove -Wall warnings.

* Fixes to warning with MPI enabled, all unused variables.

* Lots more changes to fix warnings generated by -Wall option using gcc compiler on apollo.

Note the src/eos/primitive-solver/logs.hpp file still generates warnings about
strict aliasing, but jacob assures me this is OK and will be fixed in the
future (suggested fix is move to C++20.

* Various bug fixes that have been pointed out by users over the last
year.  includes incorrectvalues for gma0/1 in IMEX3 integrator (pointed
out by Xi Lin), and errors in temperature, vorticity, and current density
calculations in derived_variables.cpp (pointed out by Michael Grehan).

* Got rid of further compiler warnings with gcc using changes suggested
by J. Fields for athena_tensor.hpp

* Fixed style errors

+++ src/eos/primitive-solver/eos_compose.cpp
@@ -195,10 +195,10 @@ void EOSCompOSE<LogPolicy>::ReadTableFromFile(std::string fname) {
     m_min_h = std::numeric_limits<Real>::max();
     // New form of bound based on properties of NQT functions and their
     // departure from 'true' log behaviour
-    int it = 0; // T = T_min is a safe assumption for the minimum enthalpy
-    for (int in = 0; in < m_nn-1; ++in) {
+    size_t it = 0; // T = T_min is a safe assumption for the minimum enthalpy
+    for (size_t in = 0; in < m_nn-1; ++in) {
       Real const nb = exp2_(host_log_nb(in));
-      for (int iy = 0; iy < m_ny-1; ++iy) {
+      for (size_t iy = 0; iy < m_ny-1; ++iy) {
         Real min_log2_e_in = Kokkos::fmin(host_table(ECLOGE,in,iy,it),
                                           host_table(ECLOGE,in,iy+1,it));
         Real max_log2_e_in = Kokkos::fmax(host_table(ECLOGE,in,iy,it),
